+++ src/status.jl
@@ -1,32 +1,20 @@
-didconverge(m::Message) = m.converged
-
 isdone(s::Symbol) = s in (:Done, :Cancelled, :Failed)
 
-iscancelled(c::RemoteChannel) = isready(c) && iscancelled(fetch(c))
-iscancelled(m::Message) = m.cancelled
+isfailed(s::Stage) = s.ex !== nothing
+isfailed(sr::StageRef) = fetch_from_owner(isfailed, sr)
+isfailed(pl::Pipeline) = any(isfailed, pl.stages)
 
-"""
-    is_pipeline_started(pl::Pipeline) -> Bool
-
-Determine whether the stages of a pipeline have been started executing.
-"""
-is_pipeline_started(pl::Pipeline) = pl.tasks !== nothing
-
-"""
-    is_pipeline_done(pl::Pipeline) -> Bool
-
-Determine whether all the stages of a pipeline have exited.
-Does not block and not throw an error, if the pipeline failed.
-"""
-is_pipeline_done(pl::Pipeline) = is_pipeline_started(pl) && all(istaskdone, pl.tasks)
+iscancelled(x::Union{Stage,Pipeline,Solution}) = x.cancelled
+iscancelled(::Message) = false
+iscancelled(::Cancellation) = true
+iscancelled(c::MessageChannel) = isready(c) && iscancelled(fetch(c))
 
 """
     is_pipeline_cancelled(pl::Pipeline) -> Bool
 
 Determine whether the pipeline had been cancelled.
 """
-is_pipeline_cancelled(pl::Pipeline) = pl.cancelled
-
+is_pipeline_cancelled(pl::Pipeline) = iscancelled(pl)
 
 """
     is_pipeline_failed(pl::Pipeline) -> Bool
@@ -34,7 +22,4 @@ is_pipeline_cancelled(pl::Pipeline) = pl.cancelled
 Determine whether some stage of a pipeline has exited because an exception was thrown.
 Does not block and not throw an error, if the pipeline failed.
 """
-function is_pipeline_failed(pl::Pipeline)
-    is_pipeline_started(pl) || return false
-    any(istaskfailed, pl.tasks)
-end
+is_pipeline_failed(pl::Pipeline) = isfailed(pl)